followup: fix import from folder restructure
