+++ src/server.tsx
@@ -10,7 +10,8 @@ import {Provider} from "react-redux";
 import {RouterProvider} from "react-router5";
 import * as favicon from "serve-favicon";
 import {config as appConfig} from "../config";
-import {App, Html} from "./app/containers";
+import {App} from "./app/containers/App";
+import {Html} from "./app/containers/Html";
 import {LanguageHelper} from "./app/helpers/LanguageHelper";
 import {configureStore} from "./app/redux/configureStore";
 import {IStore} from "./app/redux/IStore";
